Move calculatePremium tests to requestAllocation. Use divCeil in helper functions
