+++ contracts/modules/staking/StakingPool.sol
@@ -15,6 +15,7 @@ import "../../libraries/Math.sol";
 import "../../libraries/UncheckedMath.sol";
 import "../../libraries/SafeUintCast.sol";
 import "./StakingTypesLib.sol";
+import "hardhat/console.sol";
 
 // total stake = active stake + expired stake
 // total capacity = active stake * global capacity factor
@@ -1527,6 +1528,7 @@ contract StakingPool is IStakingPool, ERC721 {
     return (premiumPerYear * period / 365 days, product);
   }
 
+  // TODO: will this function work if initialCapacityUsed > totalCapacity?
   function calculatePremiumPerYear(
     uint basePrice,
     uint coverAmount,
