+++ test/unit/StakingPool/setup.js
@@ -87,6 +87,7 @@ async function setup() {
     SURGE_THRESHOLD_DENOMINATOR,
     SURGE_THRESHOLD_RATIO,
     NXM_PER_ALLOCATION_UNIT,
+    ALLOCATION_UNITS_PER_NXM,
     GLOBAL_MIN_PRICE_RATIO,
     GLOBAL_CAPACITY_RATIO,
     INITIAL_PRICE_DENOMINATOR,
@@ -99,6 +100,7 @@ async function setup() {
     stakingPool.SURGE_THRESHOLD_DENOMINATOR(),
     stakingPool.SURGE_THRESHOLD_RATIO(),
     stakingPool.NXM_PER_ALLOCATION_UNIT(),
+    stakingPool.ALLOCATION_UNITS_PER_NXM(),
     cover.GLOBAL_MIN_PRICE_RATIO(),
     cover.globalCapacityRatio(),
     stakingPool.INITIAL_PRICE_DENOMINATOR(),
@@ -122,6 +124,7 @@ async function setup() {
     SURGE_THRESHOLD_DENOMINATOR,
     SURGE_THRESHOLD_RATIO,
     NXM_PER_ALLOCATION_UNIT,
+    ALLOCATION_UNITS_PER_NXM,
     INITIAL_PRICE_DENOMINATOR,
   };
 }
